kaznu-rag/requirements.txt
beautifulsoup4==4.13.4
chardet==5.2.0
fastapi==0.115.12
lxml==5.4.0
pymupdf==1.26.0
python-dotenv==1.1.0
qdrant-client==1.14.2
requests==2.32.3
sentence-transformers==4.1.0
trafilatura==2.0.0
uvicorn==0.34.2